requirements.txt
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
huggingface-hub>=0.20.0
